Feat: 게시글 조회 API 연동, 프로젝트 구조 변경
- 프로젝트 구조 변경
- RetrofitInstance 싱글톤으로 관리하도록 수정
- post 관련 model, datasource, service, viewmodel 생성 및 수정
