+++ app/src/main/java/com/example/greetingcard/MainActivity.kt
@@ -7,12 +7,12 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.navigation.compose.rememberNavController
-import com.example.greetingcard.ui.theme.GreetingCardTheme
-import com.example.greetingcard.viewModel.login.kakao.KakaoViewModel
+import com.example.greetingcard.presentation.ui.theme.GreetingCardTheme
+import com.example.greetingcard.presentation.viewModel.login.kakao.KakaoViewModel
 
 class MainActivity : ComponentActivity() {
 
-    private val kakaoAuthViewModel : KakaoViewModel by viewModels()
+    private val kakaoAuthViewModel: KakaoViewModel by viewModels()
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
