+++ app/src/main/java/com/example/greetingcard/data/service/post/PostService.kt
@@ -1,0 +1,13 @@
+package com.example.greetingcard.data.service.post
+
+import com.example.greetingcard.data.model.response.Post
+import retrofit2.Response
+import retrofit2.http.GET
+
+interface PostService {
+
+    // 게시글 조회
+    @GET("/api/post")
+    suspend fun getPostList(
+    ): Response<List<Post>>
+}
