--- app/src/main/java/com/example/greetingcard/navigation/Screen.kt
@@ -1,12 +1,0 @@
-package com.example.greetingcard.navigation
-
-sealed class Screen(val route: String) {
-    object Login : Screen(route = "login")
-    object LoginFinder: Screen(route = "loginFinder/{viewType}")
-    object LoginJoin: Screen(route = "loginJoin")
-
-    object Splash : Screen(route = "splash_screen")
-    object Home: Screen(route = "home")
-    object CreatePlan: Screen(route = "create_plan")
-}
-
