+++ app/src/main/java/com/example/greetingcard/presentation/navigation/Screen.kt
@@ -1,0 +1,12 @@
+package com.example.greetingcard.presentation.navigation
+
+sealed class Screen(val route: String) {
+    object Login : Screen(route = "login")
+    object LoginFinder : Screen(route = "loginFinder/{viewType}")
+    object LoginJoin : Screen(route = "loginJoin")
+
+    object Splash : Screen(route = "splash_screen")
+    object Home : Screen(route = "home")
+    object CreatePlan : Screen(route = "create_plan")
+}
+
