+++ app/src/main/java/com/example/greetingcard/presentation/ui/login/view/CustomInputField.kt
@@ -1,0 +1,105 @@
+package com.example.greetingcard.presentation.ui.login.view
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.text.KeyboardOptions
+import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.text.Placeholder
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+
+
+@Composable
+fun CustomInputField(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholderText: String,
+    leadingIcon: ImageVector? = null,
+    trailingIcon: ImageVector? = null,
+    modifier: Modifier = Modifier
+) {
+    OutlinedTextField(
+        value = value,
+        onValueChange = onValueChange,
+        placeholder = {
+            Text(
+                text = placeholderText,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+        },
+        leadingIcon = leadingIcon?.let {
+            {
+                Icon(
+                    imageVector = leadingIcon,
+                    contentDescription = null
+                )
+            }
+        },
+        trailingIcon = trailingIcon?.let {
+            {
+                Icon(
+                    imageVector = it,
+                    contentDescription = null
+                )
+            }
+        },
+        modifier = modifier
+            .fillMaxWidth(0.9f)
+            .padding(vertical = 8.dp),
+        shape = RoundedCornerShape(8.dp)
+    )
+}
+
+
+@Composable
+fun CustomPwd(
+    value: String,
+    onValueChange: (String) -> Unit,
+    placeholderText: String,
+    leadingIcon: ImageVector? = null,
+    trailingIcon: ImageVector? = null,
+    modifier: Modifier = Modifier
+) {
+    OutlinedTextField(
+        value = value,
+        onValueChange = onValueChange,
+        placeholder = {
+            Text(
+                text = placeholderText,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis
+            )
+        },
+        leadingIcon = leadingIcon?.let {
+            {
+                Icon(
+                    imageVector = leadingIcon,
+                    contentDescription = null
+                )
+            }
+        },
+        trailingIcon = trailingIcon?.let {
+            {
+                Icon(
+                    imageVector = it,
+                    contentDescription = null
+                )
+            }
+        },
+        modifier = modifier
+            .fillMaxWidth(0.9f)
+            .padding(vertical = 8.dp),
+        shape = RoundedCornerShape(8.dp),
+        visualTransformation = PasswordVisualTransformation(),
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Password)
+    )
+}
