--- app/src/main/java/com/example/greetingcard/ui/theme/ui/home/planning/CreatePlanPage.kt
@@ -1,20 +1,0 @@
-package com.example.greetingcard.ui.theme.ui.home.planning
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.style.TextAlign
-
-@Composable
-fun CreatePlanPage() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center,
-    ) {
-        Text(text = "Create Plan Page", textAlign = TextAlign.Center)
-    }
-}
