--- app/src/main/java/com/example/greetingcard/ui/theme/ui/login/component/LoginTitle.kt
@@ -1,31 +1,0 @@
-package com.example.greetingcard.ui.theme.ui.login.component
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-
-@Composable
-fun LoginTitle(
-    title: String,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(top = 120.dp, bottom = 15.dp),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = title,
-            fontSize = 35.sp,
-            fontWeight = FontWeight.Bold)
-    }
-}
